FIX-2137 Clean up root.tsx
